chore: fix ts errors
